Improve context creation
