+++ src/cli.ts
@@ -1,14 +1,12 @@
 import * as process from 'node:process';
 
-import esbuild from 'esbuild';
 import { hideBin } from 'yargs/helpers';
 import yargs from 'yargs/yargs';
 
-import * as plugin from './plugins/index.js';
-import { createRunner } from './runner.js';
+import { createContext } from './context.js';
 import { Mode } from './types.js';
 import { dirname } from './utils/dirname.js';
-import { getMetadata } from './utils/read-pkg.js';
+import { Metadata, getMetadata } from './utils/read-pkg.js';
 import { rimraf } from './utils/rimraf.js';
 
 const { __dirname } = dirname(import.meta.url);
@@ -30,24 +28,17 @@ export function cli() {
           type: 'string',
         } as const,
       },
-      (argv) => {
-        let { project, bundler, config } = getMetadata(argv.cwd ?? process.cwd(), __dirname);
-        let service = createRunner({
-          mode: argv.mode,
-          watch: false,
-          config,
-          project,
-          bundler,
-          host: 'localhost',
-          port: 3000,
-        });
+      async function build(argv) {
+        let metadata = getMetadata(argv.cwd ?? process.cwd(), __dirname);
+        prerun(argv.mode, metadata);
 
-        service.subscribe((state) => {
-          if (state.matches('success')) process.exit(0);
-          if (state.matches('error')) process.exit(1);
-        });
+        let context = await createContext({ ...argv, ...metadata });
 
-        service.start();
+        try {
+          await context.rebuild();
+        } finally {
+          await context.dispose();
+        }
       },
     )
     .command(
@@ -75,13 +66,11 @@ export function cli() {
           type: 'string',
         } as const,
       },
-      async (argv) => {
+      async function dev(argv) {
         let metadata = getMetadata(argv.cwd ?? process.cwd(), __dirname);
+        prerun(argv.mode, metadata);
 
-        process.env.NODE_ENV = process.env.NODE_ENV || argv.mode === 'dev' ? 'development' : 'production';
-        rimraf(metadata.project.paths.absolute(metadata.config.outdir));
-
-        let context = await createContext(argv, metadata);
+        let context = await createContext({ ...argv, ...metadata });
 
         try {
           await context.watch();
@@ -98,72 +87,7 @@ export function cli() {
     .parse();
 }
 
-type Argv = {
-  mode?: Mode;
-  cwd?: string;
-  host?: string;
-  port?: number;
-};
-
-function createContext(argv: Argv, options: ReturnType<typeof getMetadata>) {
-  let pluginOptions = {
-    mode: 'prod' as const,
-    cwd: process.cwd(),
-    host: 'localhost',
-    port: 3000,
-    watch: argv.mode === 'dev',
-    ...options,
-    ...argv,
-  };
-
-  let plugins = [
-    plugin.reactFactory(pluginOptions),
-    plugin.define(pluginOptions),
-    plugin.externals(pluginOptions),
-    plugin.translations(pluginOptions),
-    plugin.postcss(pluginOptions),
-    plugin.assetLoader(pluginOptions),
-  ];
-
-  if (argv.mode === 'prod') {
-    plugins.unshift(plugin.nomodule(pluginOptions));
-  }
-
-  plugins.push({
-    name: 'test',
-    setup(build) {
-      build.onEnd(() => {
-        console.log('build done');
-      });
-    },
-  });
-
-  let entryNames = argv.mode === 'prod' ? '[dir]/[name].[hash]' : '[dir]/[name]';
-
-  return esbuild.context({
-    entryPoints: options.config.entryPoints,
-    outdir: options.project.paths.absolute(options.config.outdir),
-    entryNames,
-    bundle: true,
-    format: 'esm',
-    platform: 'browser',
-    target: 'es2020',
-    write: true,
-    metafile: true,
-
-    loader: {
-      '.ttf': 'file',
-      '.eot': 'file',
-      '.woff': 'file',
-      '.woff2': 'file',
-    },
-
-    sourcemap: options.config.sourcemap || argv.mode === 'dev',
-    minify: argv.mode === 'prod',
-
-    absWorkingDir: options.project.paths.root,
-    logLevel: 'silent',
-
-    plugins,
-  });
+function prerun(mode: Mode, metadata: Metadata) {
+  process.env.NODE_ENV = process.env.NODE_ENV || mode === 'dev' ? 'development' : 'production';
+  rimraf(metadata.project.paths.absolute(metadata.config.outdir));
 }
